update to breaking changes from predicate coin input type
