+++ fuel-core-interfaces/src/db.rs
@@ -4,6 +4,7 @@ use fuel_vm::prelude::InterpreterError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
+#[non_exhaustive]
 pub enum Error {
     #[error("error performing binary serialization")]
     Codec,
@@ -180,15 +181,13 @@ pub mod helpers {
                 receipts_root: Default::default(),
                 script,
                 script_data: vec![],
-                inputs: vec![Input::Coin {
+                inputs: vec![Input::CoinSigned {
                     utxo_id: UtxoId::new(*TX_ID_DB1, 0),
                     owner: Address::default(),
                     amount: 100,
                     asset_id: Default::default(),
                     witness_index: 0,
                     maturity: 0,
-                    predicate: vec![],
-                    predicate_data: vec![],
                 }],
                 outputs: vec![
                     Output::Coin {
@@ -221,15 +220,13 @@ pub mod helpers {
                 receipts_root: Default::default(),
                 script,
                 script_data: vec![],
-                inputs: vec![Input::Coin {
+                inputs: vec![Input::CoinSigned {
                     utxo_id: UtxoId::new(*TX_ID_DB1, 0),
                     owner: Address::default(),
                     amount: 100,
                     asset_id: Default::default(),
                     witness_index: 0,
                     maturity: 0,
-                    predicate: vec![],
-                    predicate_data: vec![],
                 }],
                 outputs: vec![Output::ContractCreated {
                     contract_id: *CONTRACT_ID1,
@@ -255,15 +252,13 @@ pub mod helpers {
                 receipts_root: Default::default(),
                 script,
                 script_data: vec![],
-                inputs: vec![Input::Coin {
+                inputs: vec![Input::CoinSigned {
                     utxo_id: UtxoId::new(*TX_ID1, 0),
                     owner: Address::default(),
                     amount: 100,
                     asset_id: Default::default(),
                     witness_index: 0,
                     maturity: 0,
-                    predicate: vec![],
-                    predicate_data: vec![],
                 }],
                 outputs: vec![Output::Coin {
                     amount: 100,
@@ -290,15 +285,13 @@ pub mod helpers {
                 receipts_root: Default::default(),
                 script,
                 script_data: vec![],
-                inputs: vec![Input::Coin {
+                inputs: vec![Input::CoinSigned {
                     utxo_id: UtxoId::new(*TX_ID1, 0),
                     owner: Address::default(),
                     amount: 100,
                     asset_id: Default::default(),
                     witness_index: 0,
                     maturity: 0,
-                    predicate: vec![],
-                    predicate_data: vec![],
                 }],
                 outputs: vec![
                     Output::Coin {
@@ -332,15 +325,13 @@ pub mod helpers {
                 receipts_root: Default::default(),
                 script,
                 script_data: vec![],
-                inputs: vec![Input::Coin {
+                inputs: vec![Input::CoinSigned {
                     utxo_id: UtxoId::new(*TX_ID_DB1, 0),
                     owner: Address::default(),
                     amount: 100,
                     asset_id: Default::default(),
                     witness_index: 0,
                     maturity: 0,
-                    predicate: vec![],
-                    predicate_data: vec![],
                 }],
                 outputs: vec![Output::Coin {
                     amount: 100,
@@ -368,15 +359,13 @@ pub mod helpers {
                 receipts_root: Default::default(),
                 script,
                 script_data: vec![],
-                inputs: vec![Input::Coin {
+                inputs: vec![Input::CoinSigned {
                     utxo_id: UtxoId::new(*TX_ID_DB2, 0),
                     owner: Address::default(),
                     amount: 200,
                     asset_id: Default::default(),
                     witness_index: 0,
                     maturity: 0,
-                    predicate: vec![],
-                    predicate_data: vec![],
                 }],
                 outputs: vec![Output::Coin {
                     amount: 100,
