+++ fuel-relayer/Cargo.toml
@@ -20,7 +20,7 @@ ethers-providers = { git = "https://github.com/rakita/ethers-rs.git", branch = "
 ] }
 features = "0.10"
 fuel-core-interfaces = { path = "../fuel-core-interfaces", package = "fuel-core-interfaces", version = "0.7.1" }
-fuel-tx = { version = "0.10", features = ["serde"] }
+fuel-tx = { version = "0.11", features = ["serde"] }
 fuel-types = { version = "0.5", features = ["serde"] }
 futures = "0.3"
 lazy_static = "1.4"
